Chore: atomic design에 맞게 폴더 구조 수정
