+++ Frontend/src/components/organisms/HeaderBar-MenuDropdown/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 
 import styled from 'styled-components';
-import MenuLinkBtn from 'components/MenuLinkBtn';
+import MenuLinkBtn from 'components/molecules/MenuLinkBtn';
 
 const Main = styled.div<{clicked: boolean}>`
   position: absolute;
