+++ Frontend/src/components/organisms/PleaseScroll/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import * as S from './style';
+
+function App() {
+  return (
+    <S.Main>
+      <div>PKIOP가 궁금하시면? Scroll👇 </div>
+    </S.Main>
+  );
+}
+
+export default App;
