+++ Frontend/src/components/organisms/PleaseScroll/style.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const Main = styled.div`
+  height: ${({ theme }) => theme.pleaseSlideHeight}rem;
+  color: yellow;
+  background-color: ${({ theme }) => theme.black};
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  div {
+
+  }
+`;
+export default {};
