+++ Frontend/src/components/organisms/TextSlide/index.tsx
