+++ Frontend/src/components/organisms/TextSlide/text.tsx
