+++ Frontend/src/pages/About/style.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Main = styled.div`
+`;
+
+export default {};
