+++ Frontend/src/pages/PostComponent/index.tsx
