+++ Frontend/src/styles/global-styles.ts
@@ -20,7 +20,6 @@ export const GlobalStyle = createGlobalStyle`
     text-decoration: none;
     color: inherit;
   }
-
 `;
 
 export default {};
